+++ src/main/resources/com/example/programacion4proyectofinal/JSON/AdminUser.json
@@ -1,0 +1,4 @@
+{
+  "username": "progra",
+  "password": "prograiv"
+}
